requirements.txt
google-generativeai==0.8.2
pytest==8.3.3
openai==0.28.0
